% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVA.R
\name{AoD.biplot}
\alias{AoD.biplot}
\title{AoD biplot}
\usage{
\method{AoD}{biplot}(
  bp,
  Dmat = NULL,
  dist.func = NULL,
  dim.biplot = c(2, 1, 3),
  e.vects = 1:ncol(bp$X),
  classes = bp$classes,
  weighted = c("unweighted", "weighted"),
  show.class.means = TRUE,
  axes = "regression",
  ...
)
}
\arguments{
\item{bp}{an object of class \code{biplot} obtained from preceding function \code{biplot()}.}

\item{Dmat}{nxn matrix of Euclidean embeddable distances between samples}

\item{dist.func}{function to compute Euclidean embeddable distances between samples. The
default NULL computes Euclidean distance.}

\item{dim.biplot}{dimension of the biplot. Only values 1, 2 and 3 are accepted, with default \code{2}.}

\item{e.vects}{e.vects which eigenvectors (canonical variates) to extract, with default \code{1:dim.biplot}.}

\item{classes}{classes vector of the same length as the number of rows in the data matrix
with the class indicator for the samples.}

\item{weighted}{the default is "unweighted" where each class receives equal weight, alternatively "weighted" use class sizes as weights.}

\item{show.class.means}{logical, indicating whether to plot the class means on the biplot.}

\item{axes}{type of biplot axes, currently only regression axes are implemented}

\item{...}{more arguments to \code{dist.func}}
}
\value{
an object of class biplot.
}
\description{
Computes Analysis of Distance biplot
}
\examples{
biplot(iris) |> AoD(classes = iris[,5]) |> plot()

}
