% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{biplot.spline.axis.control}
\alias{biplot.spline.axis.control}
\title{Spline Axis Control}
\usage{
biplot.spline.axis.control(
  tau = 0.5,
  nmu = 100,
  u = 2,
  v = 3,
  lambda = 0,
  smallsigma = 0.01,
  bigsigma = smallsigma * 10,
  gamma = 250,
  bigsigmaactivate = floor(gamma * 0.1),
  eps = 0.01,
  tiny = 10^-30,
  itmax = 10000,
  ftol = 1.5 * (.Machine$double.eps)
)
}
\arguments{
\item{tau}{}

\item{nmu}{}

\item{u}{}

\item{v}{}

\item{lambda}{}

\item{smallsigma}{}

\item{bigsigma}{}

\item{gamma}{}

\item{bigsigmaactivate}{}

\item{eps}{}

\item{tiny}{}

\item{itmax}{}

\item{ftol}{}
}
\description{
Spline Axis Control
}
