% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{biplot.spline.axis}
\alias{biplot.spline.axis}
\title{calibrate spline based axes on biplots}
\usage{
biplot.spline.axis(j, X, Y, means, sd, n.int, spline.control, dmeth = 0, ...)
}
\arguments{
\item{j}{Index of the axis to be calibrated in the data}

\item{X}{Coordinates of the samples on the biplot space}

\item{means}{column means of the raw data}

\item{sd}{Column standard deviations of the data}

\item{n.int}{the number of tick marks per axis}

\item{spline.control}{control variables for optimisation. See biplotEZ:::biplot.spline.axis.control()}

\item{dmeth}{Argument unused}

\item{...}{additional arguments}

\item{Ytilde}{Raw data used to construct the biplot}
}
\description{
calibrate spline based axes on biplots
}
