% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATPCA.R
\name{CatPCA.biplot}
\alias{CatPCA.biplot}
\title{Calculate elements for the CatPCA biplot}
\usage{
\method{CatPCA}{biplot}(
  bp,
  cat.type = NULL,
  dim.biplot = c(2, 1, 3),
  e.vects = 1:ncol(bp$X),
  group.aes = NULL,
  show.class.means = FALSE,
  epsilon = 1e-06
)
}
\arguments{
\item{bp}{an object of class \code{biplot} obtained from preceding function \code{biplot()}.}

\item{cat.type}{a vector indicating whether each categorical variables is \code{nominal} or \code{ordinal}.}

\item{dim.biplot}{the dimension of the biplot. Only values \code{1}, \code{2} and \code{3} are accepted, with default \code{2}.}

\item{e.vects}{the vector indicating which eigenvectors (principal components) should be plotted in the biplot, with default \code{1:dim.biplot}.}

\item{group.aes}{a vector of the same length as the number of rows in the data matrix
for differentiated aesthetics for samples.}

\item{show.class.means}{a logical value indicating whether group means should be plotted in the biplot.}

\item{epsilon}{convergence criteria for iterative algorithm with defaults 1e-6.}
}
\value{
an object of class \code{CatPCA}, inherits from classes \code{PCA} and \code{biplot}.
}
\description{
This function performs calculations for the construction of a CatPCA biplot.
}
\examples{
biplot(iris) |> CatPCA()
# create a CatPCA biplot
biplot(data = iris) |> CatPCA() |> plot()

}
