% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATPCA.R
\name{CATPCA}
\alias{CATPCA}
\title{Categorical Principal Component Analysis}
\usage{
CATPCA(bp, dim.biplot = c(2, 1, 3), e.vects = 1:ncol(bp$X),
group.aes = NULL, show.class.means = FALSE)
}
\arguments{
\item{bp}{an object of class \code{biplot} obtained from preceding function \code{biplot()}.}

\item{dim.biplot}{dimension of the biplot. Only values 1, 2 and 3 are accepted, with default \code{2}.}

\item{e.vects}{which eigenvectors (principal components) to extract, with default \code{1:dim.biplot}.}

\item{group.aes}{vector of the same length as the number of rows in the data matrix
for differentiated aesthetics for samples.}

\item{show.class.means}{logical, indicating whether group means should be plotted in the biplot.}
}
\value{
an object of class \code{biplot}
}
\description{
Categorical Principal Component Analysis
}
\examples{
CATPCA (iris)
}
