% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{classify}
\alias{classify}
\title{Classify samples into classes}
\usage{
classify(
  bp,
  classify.regions = TRUE,
  col = ez.col,
  opacity = 0.4,
  borders = FALSE
)
}
\arguments{
\item{bp}{an object of class \code{biplot}}

\item{classify.regions}{a logical value indicating whether classifications regions are
shown in the biplot, with default \code{TRUE}.}

\item{col}{the colours of the classification regions}

\item{opacity}{the opacity levels of the classification regions}

\item{borders}{the border colours of the classification regions}
}
\value{
An object of class \code{biplot}
}
\description{
Classify samples into classes
}
\examples{
biplot(iris[,1:4],classes = iris[,5]) |> CVA() |> axes(col="black") |> 
  classify(col=c("red","blue","orange"),opacity=0.1) |> plot()
}
