% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{classmeans}
\alias{classmeans}
\title{Formatting of classmeans in biplot}
\usage{
classmeans(
  bp,
  col = c("blue", "green", "gold", "cyan", "magenta", "black", "red", "#33A02C",
    "purple", "salmon"),
  pch = 3,
  cex = 1,
  label = F,
  label.cex = 0.75,
  label.side = "bottom",
  connected = F,
  alpha = 1
)
}
\arguments{
\item{alpha}{}
}
\description{
Formatting of classmeans in biplot
}
