% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{CLPs}
\alias{CLPs}
\title{Aesthetics for category level points}
\usage{
CLPs (bp,  which = 1:ncol(bp$Xcat), col = "black", cex = 0.6)
}
\arguments{
\item{bp}{an object of class \code{biplot}.}

\item{which}{vector of which variables' CLPs to display, with default \code{ncol(Xcat)}.}

\item{col}{list of CLP colours, with default \code{black}.}

\item{cex}{list of CLP character expansions, with default \code{0.6}.}
}
\value{
A list with the following components is available:
\item{which}{which variables' CLPs to display.}
\item{col}{colour of the CLPs.}
\item{cex}{expansion of the plotting character of the CLPs.}
}
\description{
This function allows formatting changes to CLPs.
}
\examples{
mtdf <- as.data.frame(mtcars)
mtdf$cyl <- factor(mtdf$cyl)
mtdf$vs <- factor(mtdf$vs)
mtdf$am <- factor(mtdf$am)
mtdf$gear <- factor(mtdf$gear)
mtdf$carb <- factor(mtdf$carb)
biplot(mtdf[,-11], scaled = TRUE) |> AoD(classes = mtdf[,11]) |> 
CLPs(col = list(rep("olivedrab",3), rep("orange",2),
                rep("coral",2), rep("brown",3))) |> 
plot()

}
