% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{concentration.ellipse}
\alias{concentration.ellipse}
\title{Concentration ellipses}
\usage{
concentration.ellipse(
  bp,
  df = 2,
  kappa = NULL,
  which = NULL,
  alpha = 0.95,
  col = ez.col,
  lty = 1,
  lwd = 1,
  alpha.transparency = 0.5
)
}
\arguments{
\item{bp}{object of class biplot}

\item{df}{degrees of freedom}

\item{kappa}{value to construct k-ellipse}

\item{which}{which group to select for ellipse construction}

\item{alpha}{size of alpha bag}

\item{col}{colour of ellipse}

\item{lty}{line type of ellipse}

\item{lwd}{line width of ellipse}

\item{alpha.transparency}{opacity}
}
\value{
an object of class biplot
}
\description{
Concentration ellipses
}
\examples{
biplot (iris[,1:4]) |> PCA(group.aes=iris[,5]) |> concentration.ellipse(kappa=2) |> plot()
}
