% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVA.R
\name{CVA.2class}
\alias{CVA.2class}
\title{Construct additional dimensions when the dimension of the canonical space is lower than the dimension of the biplot}
\usage{
\method{CVA}{`2class`}(bp, G, W, Mmat, low.dim, K, e.vects)
}
\arguments{
\item{bp}{an object of class \code{biplot}.}

\item{G}{the indicator matrix defining membership of the classes.}

\item{W}{the within class sums of squares and cross products matrix.}

\item{Mmat}{eigenvector matrix from CVA}

\item{low.dim}{if the dimension of the canonical space is smaller than \code{dim.biplot}, the method to use to construct
additional dimension(s). Currently two options are implemented: \code{sample.opt} maximises the sample predictivity
of the individual samples in the biplot and \code{Bhattacharyya.dist} is based on the decomposition of the
Bhattacharyya distance into a component for the sample means and a component for the dissimilarity
between the sample covariance matrices.}

\item{K}{the dimension of the canonical space}

\item{e.vects}{the vector indicating which canonical variates are plotted in the biplot, with default \code{1:dim.biplot}}
}
\value{
A list with three components:
\item{Mr}{pxr matrix for the transformation to the canonical space.}
\item{Mrr}{rxp matrix for prediction from the canonical space.}
\item{Lmat}{pxp matrix for transformation to the canonical space.}
}
\description{
Construct additional dimensions when the dimension of the canonical space is lower than the dimension of the biplot
}
