% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouds.R
\name{density1D}
\alias{density1D}
\title{Kernel density estimates of 1D samples}
\usage{
density1D(
  bp,
  which = NULL,
  bw = "nrd0",
  kernel = "gaussian",
  col = ez.col,
  lwd = 1.5,
  legend.mar = c(2, 5, 0, 5),
  ...
)
}
\arguments{
\item{bp}{an object of class \code{biplot}}

\item{which}{the selection of groups or classes to be fitted with densitiy estimates}

\item{bw}{the smoothing bandwidth to be used.See \code{?density}
for more details}

\item{kernel}{a character string giving the smoothig kernel to be used. See \code{?density}
for more details}

\item{col}{vector of colours for the density estimates}

\item{lwd}{vector  of line widths for the density estimates}

\item{legend.mar}{margin width for the legend}

\item{...}{additional arguments to be passed to \code{density}}
}
\description{
Kernel density estimates of 1D samples
}
