% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{extended.matching.coefficient}
\alias{extended.matching.coefficient}
\title{Extended matching coefficient}
\usage{
extended.matching.coefficient(X)
}
\arguments{
\item{X}{a data frame containing the categorical variables used for computing the EMC distance}
}
\value{
a dist object
}
\description{
Extended matching coefficient
}
\examples{
mtdf <- as.data.frame(mtcars)
mtdf$cyl <- factor(mtdf$cyl)
mtdf$vs <- factor(mtdf$vs)
mtdf$am <- factor(mtdf$am)
mtdf$gear <- factor(mtdf$gear)
mtdf$carb <- factor(mtdf$carb)
extended.matching.coefficient(mtdf[,8:11])

}
