% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{PCA.biplot}
\alias{PCA.biplot}
\title{PCA biplot}
\usage{
\method{PCA}{biplot}(
  bp,
  dim.biplot = c(2, 1, 3),
  e.vects = 1:ncol(bp$X),
  group.aes = NULL,
  correlation.biplot = FALSE,
  ...
)
}
\arguments{
\item{bp}{object of class PCA}

\item{dim.biplot}{dimension of the biplot. Only values 1, 2 and 3 are accepted, with default 2}

\item{e.vects}{which eigevectors (principal components) to extract, defaults to \code{1:dim.biplot}}

\item{group.aes}{optional vector of the same length as the number of rows in the data matrix
for differentiated aesthetics for samples}

\item{correlation.biplot}{defaults to FALSE. If FALSE, the distances between sample points are
optimally approximated in the biplot. If TRUE, the correlations between
variables are optimally approximated by the cosine of the angles between
axes. See Gabriel (1971) The biplot graphic display of matrices with application
to principal component analysis. Biometrika, 58(3), pp.453-467.}

\item{...}{additional arguments}
}
\value{
an object of class ??
}
\description{
Performs calculations for a PCA biplot
}
\examples{
biplot(iris[,1:4]) |> PCA()

}
