% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.biplot}
\alias{plot.biplot}
\title{Generic Plotting function of objects of class biplot}
\usage{
\method{plot}{biplot}(
  x,
  engine = c("ggplot2", "base"),
  exp.factor = 1.2,
  axis.predictivity = NULL,
  sample.predictivity = NULL,
  zoom = FALSE,
  add = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{biplot}.}

\item{engine}{whether to plot in base R or \code{ggplot2}.}

\item{exp.factor}{a numeric value with default axes of the biplot. Larger values are specified for zooming out with respect to sample points in the biplot display and smaller values are specified for zooming in with respect to sample points in the biplot display.}

\item{axis.predictivity}{either a logical or a numeric value between \code{0} and \code{1}. If it is a numeric value, this value is used as threshold so that only axes with axis predictivity larger than the threshold is displayed. If \code{axis.predictivity = TRUE}, the axis colour is 'diluted' in proportion with the axis predictivity.}

\item{sample.predictivity}{either a logical or a numeric value between 0 and 1. If it is a numeric value, this value is used as threshold so that only samples with sample predictivity larger than the threshold is displayed. If \code{sample.predictivity = TRUE}, the sample size is shrinked in proportion with the sample predictivity.}

\item{zoom}{a logical value allowing the user to select an area to zoom into.}

\item{add}{a logical value allowing the user to add the biplot to a current plot. If \code{add = TRUE} the argument \code{zoom} is inactive.}

\item{xlim}{the horizontal limits of the plot.}

\item{ylim}{the vertical limits of the plot.}

\item{...}{additional arguments.}
}
\value{
An object of class \code{biplot}.
}
\description{
Generic Plotting function of objects of class biplot
}
\examples{
biplot (iris[,1:4]) |> PCA() |> plot()
}
