% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot3D}
\alias{plot3D}
\title{Generic Plotting function of objects of class biplot in three dimensions}
\usage{
plot3D(bp, exp.factor = 1.2, ...)
}
\arguments{
\item{bp}{an object of class \code{biplot}}

\item{exp.factor}{factor to expand plotting area beyond samples.}

\item{...}{more arguments}
}
\value{
an object of class \code{biplot}
}
\description{
Generic Plotting function of objects of class biplot in three dimensions
}
\examples{
biplot(data = iris) |> PCA(dim.biplot = 3) |> plot3D()
}
