% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{print.biplot}
\alias{print.biplot}
\title{Generic print function for objects of class biplot}
\usage{
\method{print}{biplot}(x, ...)
}
\arguments{
\item{x}{an object of class \code{biplot}.}

\item{...}{additional arguments.}
}
\value{
This function will not produce a return value, it is called for side effects.
}
\description{
This function is used to print output when the biplot object is created.
}
\examples{
out <- biplot (iris[,1:4]) |> PCA()
out
}
