% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{print.CA}
\alias{print.CA}
\title{Generic print function of objects of class CA}
\usage{
\method{print}{CA}(x, ...)
}
\arguments{
\item{x}{an object of class \code{CA}}

\item{...}{additional arguments}
}
\value{
no return value, called for side effects
}
\description{
Generic print function of objects of class CA
}
\examples{
out <- biplot(HairEyeColor[,,1], center=FALSE) |> CA()
out

}
