% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS.R
\name{regress.biplot}
\alias{regress.biplot}
\title{Regression biplot}
\usage{
\method{regress}{biplot}(bp, Z, group.aes = NULL, show.group.means = TRUE)
}
\arguments{
\item{bp}{an object of class \code{biplot} obtained from preceding function \code{biplot()}.}

\item{Z}{the matrix of coordinates of the samples}

\item{group.aes}{vector of the same length as the number of rows in the data matrix
for differentiated aesthetics for samples.}

\item{show.group.means}{logical, indicating whether group means should be plotted in the biplot.}
}
\value{
an object of class \code{biplot}.
}
\description{
Computes regression biplot axes
}
\examples{
biplot(iris) |> regress(Z = cmdscale(dist(iris[,1:4]))) |> plot()

}
