% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{rotate}
\alias{rotate}
\title{Rotate}
\usage{
rotate(bp, rotate.degrees = 0)
}
\arguments{
\item{bp}{object of class \code{biplot}}

\item{rotate.degrees}{Default is \code{0}. Positive value results in anti-clockwise rotation and negative value in clockwise rotation.}
}
\value{
An object of class \code{biplot}.
}
\description{
Rotate
}
\examples{
biplot(iris[,1:4],group.aes = iris[,5]) |> PCA() |> rotate(200) |> plot()
}
