% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sqrtManhattan}
\alias{sqrtManhattan}
\title{Computes the square root of the Manhattan distance
An example of a Euclidean embeddable distance metric}
\usage{
sqrtManhattan(X)
}
\arguments{
\item{X}{matrix of samples x variables for computation of samples x samples distance matrix}
}
\value{
a dist object
}
\description{
Computes the square root of the Manhattan distance
An example of a Euclidean embeddable distance metric
}
\examples{
sqrtManhattan(iris[,1:4])
}
