% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{summary.biplot}
\alias{summary.biplot}
\title{Generic summary function of objects of class biplot}
\usage{
\method{summary}{biplot}(
  object,
  adequacy = TRUE,
  axis.predictivity = TRUE,
  sample.predictivity = TRUE,
  class.predictivity = TRUE,
  within.class.axis.predictivity = TRUE,
  within.class.sample.predictivity = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class \code{biplot}.}

\item{adequacy}{logical, indicating whether variable adequacies should be reported.}

\item{axis.predictivity}{logical, indicating whether axis predictivities should be reported.}

\item{sample.predictivity}{logical, indicating whether sample predictivities should be reported.}

\item{class.predictivity}{logical, indicating whether class predictivities should be reported
(only applicable to objects of class \code{CVA}).}

\item{within.class.axis.predictivity}{logical, indicating whether within class axis predictivity
should be reported(only applicable to objects
of class \code{CVA}).}

\item{within.class.sample.predictivity}{logical, indicating whether within class sample predictivity
should be reported(only applicable to objects
of class \code{CVA}).}

\item{...}{additional arguments.}
}
\value{
no return value, called for side effects.
}
\description{
Generic summary function of objects of class biplot
}
\examples{
out <- biplot (iris[,1:4]) |> PCA() |> fit.measures()
summary(out)
}
