% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ties}
\alias{ties}
\title{Applies primary or secondary approach to ties}
\usage{
ties(d, delta, method = c("primary", "secondary"))
}
\arguments{
\item{d}{observed dissimilarities}

\item{delta}{given dissimilarities to match (order)}

\item{method}{either \code{primary} or \code{secondary} approach to ties}
}
\value{
a vector of ordered values
}
\description{
Applies primary or secondary approach to ties
}
