% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{alpha.bags}
\alias{alpha.bags}
\title{Create alpha bags}
\usage{
alpha.bags(
  bp,
  alpha = 0.95,
  which = NULL,
  col = ez.col,
  lty = 1,
  lwd = 1,
  max = 2500
)
}
\arguments{
\item{bp}{object of class biplot}

\item{alpha}{value between 0 and 1 to determine coverage of bag}

\item{which}{which groups or classes to be fitted with alpha bags}

\item{col}{vector of colours for the alpha bags}

\item{lty}{vector of line types for the alpha bags}

\item{lwd}{vector of line widths for the alpha bags}

\item{max}{maximum number of samples to include in alpha-bag calculations. Defaults to 2500, if
more samples are in the group, a random sample of size max is taken for the computations}
}
\value{
an object of class biplot
}
\description{
Create alpha bags
}
\examples{
biplot (iris[,1:4]) |> PCA(group.aes=iris[,5]) |> alpha.bags(alpha=0.95) |> plot()

}
