% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{axes}
\alias{axes}
\title{Aesthetics for biplot axes}
\usage{
axes(
  bp,
  X.names = colnames(bp$X),
  which = 1:bp$p,
  col = grey(0.7),
  lwd = 1,
  lty = 1,
  label.dir = "Orthog",
  label.col = col,
  label.cex = 0.75,
  label.dist = 0,
  ticks = 5,
  tick.col = col,
  tick.size = 1,
  tick.label = T,
  tick.label.col = tick.col,
  tick.label.cex = 0.6,
  tick.label.side = "left",
  tick.label.offset = 0.5,
  tick.label.pos = 1,
  predict.col = col,
  predict.lwd = lwd,
  predict.lty = lty,
  ax.names = X.names,
  orthogx = 0,
  orthogy = 0
)
}
\arguments{
\item{bp}{object of class biplot}

\item{X.names}{data column names for which axes should be constructed}

\item{which}{column numbers for which axes should be constructed}

\item{col}{axis colours}

\item{lwd}{axis line widths}

\item{lty}{axis line types}

\item{label.dir}{direction of the axis label, one of "Orthog", "Hor" or "Paral"}

\item{label.col}{colour of the axis label}

\item{label.cex}{character expansion of the axis label}

\item{label.dist}{distance of the axis label from the plot}

\item{ticks}{indicator of the number of tick marks on each axis}

\item{tick.col}{colour of tickmarks}

\item{tick.size}{size of tickmarks}

\item{tick.label}{logical whether tick marks should be labelled}

\item{tick.label.col}{colour of tick mark labels}

\item{tick.label.cex}{character expansion of tick mark labels}

\item{tick.label.side}{on which side of tick marks should label appear}

\item{tick.label.offset}{offset for position of tick mark labels}

\item{tick.label.pos}{pos argument for \code{text} function for tick mark labels}

\item{predict.col}{colour of lines showing orthogonal projection onto axes}

\item{predict.lwd}{line width of lines showing orthogonal projection onto axes}

\item{predict.lty}{line type of lines showing orthogonal projection onto axes}

\item{ax.names}{user specified alternative names for variables}

\item{orthogx}{horisontal translation}

\item{orthogy}{vertical translation}
}
\value{
object of class biplot
}
\description{
Aesthetics for biplot axes
}
\examples{
biplot(iris[,1:4]) |> PCA() |> axes(col="purple") |> plot()

}
