% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{biplot}
\alias{biplot}
\title{Create a new biplot}
\usage{
biplot(X, center = TRUE, scaled = FALSE)
}
\arguments{
\item{X}{Default dataset to use for the biplot. Dataframe or matrix.}

\item{center}{Logical value indicating whether to column centre the dataset.}

\item{scaled}{Logical value indicating wehther to standardise to unit
standard deviation.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   returns an object of class biplot
           An object of class "biplot" is a list containing at least the
           following components
           X dataset ready for biplot calculations
           raw.X original data
           scaled center ...
           means vector of column means
           sd vector of column standard deviations
}\if{html}{\out{</div>}}
}
\description{
Create a new biplot
}
