% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{biplot}
\alias{biplot}
\title{Create a new biplot}
\usage{
biplot(data, group.aes = NULL, center = TRUE, scaled = FALSE, Title = NULL)
}
\arguments{
\item{data}{A dataframe or matrix containing all variables the user wants to analyse}

\item{group.aes}{Variable from the data to be used as a grouping variable}

\item{center}{TRUE or FALSE}

\item{scaled}{TRUE or FALSE}

\item{Title}{Title}
}
\value{
a list with the following components
\itemize{
\item{X}{Matrix of the centered and scaled numeric variables}
\item{raw.X}{Original data}
\item{center}{TRUE or FALSE, should the numeric data be centred?}
\item{scaled}{TRUE or FALSE, should the numeric data be scaled?}
\item{means}{Vector of means for each numeric variable}
\item{sd}{Vector of standard deviations for each numeric variable}
\item{group.aes}{Vector of category levels for the grouping variable. This is to be used for color, pch and cex specifications}
}
}
\description{
Create a new biplot
}
\examples{
biplot(data = iris)
}
