% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{biplot.spline.axis}
\alias{biplot.spline.axis}
\title{Plot spline based axes on biplots}
\usage{
biplot.spline.axis(
  j,
  X,
  Ytilde,
  means,
  sd,
  n.int,
  spline.control,
  dmeth = 0,
  ...
)
}
\arguments{
\item{j}{}

\item{X}{}

\item{Ytilde}{}

\item{means}{}

\item{sd}{}

\item{n.int}{}

\item{spline.control}{}

\item{dmeth}{}

\item{...}{}
}
\description{
Plot spline based axes on biplots
}
