% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{classify}
\alias{classify}
\title{Classification of samples into classes}
\usage{
classify(
  bp,
  classify.regions = TRUE,
  col = ez.col,
  shade = 0.4,
  borders = FALSE
)
}
\arguments{
\item{bp}{object of class \code{biplot}}

\item{classify.regions}{logical indicating whether classifications regions are
shown in the biplot}

\item{col}{color of classification regions}

\item{shade}{opacity of classification regions}

\item{borders}{border colour of classification regions}
}
\value{
An object of class \code{biplot}
}
\description{
Classification of samples into classes
}
\examples{
biplot(iris[,1:4],classes = iris[,5]) |> CVA() |> axes(col="black") |> 
  classify(col=c("red","blue","orange"),shade=0.1) |> plot()
}
