% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVA.R
\name{CVA.2class}
\alias{CVA.2class}
\title{CVA extra dimensions if dim(canonical space) < dimension of the biplot}
\usage{
\method{CVA}{`2class`}(bp, G, W, Mmat, low.dim)
}
\arguments{
\item{bp}{object of class \code{biplot}}

\item{G}{indicator matrix of class membership}

\item{W}{within class covariance matrix}

\item{Mmat}{eigenvector matrix from CVA}

\item{low.dim}{if the dimension of the canonical space is smaller than \code{dim.biplot}, the method to use to construct
additional dimension(s). Currently two options are implemented: \code{sample.opt} maximises the sample predictivity
of the individual samples in the biplot and \code{Bhattacharyya.dist} is based on the decomposition of the
Bhattacharyya distance into a component for the sample means and a component for the dissimilarity
between the sample covariance matrices.}
}
\value{
A list with three components:
\item{Mr}{pxr matrix for the transformation to the canonical space.}
\item{Mrr}{rxp matrix for prediction from the canonical space.}
\item{Lmat}{pxp matrix for transformation to the canonical space.}
}
\description{
CVA extra dimensions if dim(canonical space) < dimension of the biplot
}
