% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{CVA}
\alias{CVA}
\title{Canonical Variate Analysis (CVA) method}
\usage{
CVA(bp, dim.biplot = c(2, 1, 3), e.vects = 1:ncol(bp$X),
group.aes = bp$group.aes,weightedCVA = "weighted")
}
\arguments{
\item{bp}{Object of class biplotAn object of class \code{biplot} obtained from preceding function \code{biplot()}.}

\item{dim.biplot}{Dimension of the biplot. Only values 1, 2 and 3 are accepted, with default \code{2}.}

\item{e.vects}{Which eigenvectors (canonical variates) to extract, with default \code{1:dim.biplot}.}

\item{group.aes}{Vector of the same length as the number of rows in the data matrix
for differentiated aesthetics for samples.}

\item{weightedCVA}{The default is "weighted", specifying a weighted CVA to be performed. Other possible values are "unweightedI" and "unweightedCent".}
}
\value{
Object of class CVA with the following elements:
\item{X}{Matrix of the centered and scaled numeric variables.}
\item{Xcat}{Matrix of the categorical variables.}
\item{raw.X}{Original data.}
\item{na.action}{Vector of observations that have been removed.}
\item{center}{TRUE or FALSE, as specified.}
\item{scaled}{TRUE or FALSE, as specified.}
\item{means}{Mean of each numerical variable.}
\item{sd}{Standard deviation of each numerical variable.}
\item{n}{Number of observations.}
\item{p}{Number of variables.}
\item{group.aes}{Vector of the same length as the number of rows in the data matrix for differentiated aesthetics for samples. }
\item{g.names}{Descriptive name to be used for group labels.}
\item{g}{Number of groups.}
\item{Title}{Title of the biplot to be rendered, as specified.}
\item{Z}{Matrix with each row containing the details of the point to be plotted (i.e. coordinates).}
\item{Xhat}{Predictions of the samples.}
\item{ax.one.unit}{One unit in the positive direction of each biplot axis.}
}
\description{
This function produces a list of elements to be used for CVA biplot construction.
}
\examples{
biplot(iris[,1:4],iris[,5]) |> CVA()
}
