% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{legend.type}
\alias{legend.type}
\title{Determines which legends to make}
\usage{
legend.type(bp, samples = F, means = F, bags = F, new = FALSE, ...)
}
\arguments{
\item{bp}{object of class biplot}

\item{samples}{logical, whether legend should be printed for samples}

\item{means}{logical, whether legend should be printed for means}

\item{bags}{logical, whether legend should be printed for bags}

\item{new}{logical, whether the legend should appear on a new page}

\item{...}{more arguments to be sent to \code{legend} function}
}
\value{
an object of class biplot
}
\description{
Determines which legends to make
}
\examples{
biplot (iris[,1:4], Title="Test biplot") |> PCA(group.aes = iris[,5]) |>
    legend.type(samples=TRUE) |> plot()
}
