% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{legend.type}
\alias{legend.type}
\title{Legend type}
\usage{
legend.type(bp, samples = FALSE, means = FALSE, bags = FALSE, ellipses = FALSE,
new = FALSE)
}
\arguments{
\item{bp}{An object of class \code{biplot}.}

\item{samples}{Logical argument indicating whether legend should be printed for samples, with default \code{FALSE}.}

\item{means}{Logical argument indicating whether legend should be printed for means, with default \code{FLASE}.}

\item{bags}{Logical argument indicating whether legend should be printed for bags, with default \code{FALSE}.}

\item{ellipses}{Logical argument indicating whether legend should be printed for concentration ellipses, with default \code{FALSE}.}

\item{new}{Logical argument indicating whether the legend should appear in a new window, with default \code{FALSE}.}

\item{...}{addition arguments to be sent to \code{legend()}.}
}
\value{
A list with the following components is available:
\item{samples}{TRUE or FALSE, as specified.}
\item{means}{TRUE or FALSE, as specified.}
\item{bags}{TRUE or FALSE, as specified.}
\item{ellipses}{TRUE or FALSE, as specified.}
\item{new}{TRUE or FALSE, as specified.}
}
\description{
This function enables the user to format the legend and make a required selection to display.
}
\examples{
biplot (iris[,1:4], Title="Test biplot") |> PCA(group.aes = iris[,5]) |>
    legend.type(samples=TRUE) |> plot()
}
