% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA.R
\name{MCA.biplot}
\alias{MCA.biplot}
\title{MCA biplot}
\usage{
\method{MCA}{biplot}(
  bp,
  dim.biplot = c(2, 1, 3),
  e.vects = 1:ncol(bp$Xcat),
  variant = "Indicator",
  lvl.num = TRUE
)
}
\arguments{
\item{bp}{object of class \code{biplot} obtained from preceding function \code{biplot(center = FALSE)}.}

\item{dim.biplot}{dimension of the biplot. Only values 1, 2 and 3 are accepted, with default \code{2}.}

\item{e.vects}{which eigenvectors (canonical variates) to extract, with default \code{1:dim.biplot}.}

\item{variant}{which multiple correspondence analysis variant, with default "Indicator", SVD performed on the weighted indicator matrix. \code{variant = "Burt"}, SVD performed on the weighted Burt matrix.}

\item{lvl.num}{logical value to specify the labels of the category levels, with default \code{TRUE} to indicate levels with numbers (V1, V2, etc.). Alternatively the level names are used.}
}
\value{
an object of class MCA, inherits from class biplot.
}
\description{
Performs calculations for an MCA biplot.
}
\examples{
biplot(as.data.frame(HairEyeColor), center = FALSE) |> 
MCA(lvl.num = FALSE) |>  samples(col=c("purple3","forestgreen"), 
pch=c(15,17), label = TRUE, label.offset = 1, label.col="gray37") |> 
plot()

}
