% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA.R
\name{MCA}
\alias{MCA}
\title{Multiple Correspondence Analysis (MCA) method}
\usage{
MCA(bp, dim.biplot = c(2,1,3), e.vects = 1:ncol(bp$Xcat), 
       variant = "Indicator", lvl.num = TRUE)
}
\arguments{
\item{bp}{object of class \code{biplot} obtained from preceding function \code{biplot(center = FALSE)}.}

\item{dim.biplot}{dimension of the biplot. Only values 1, 2 and 3 are accepted, with default \code{2}.}

\item{e.vects}{which eigenvectors (canonical variates) to extract, with default \code{1:dim.biplot}.}

\item{variant}{which multiple correspondence analysis variant, with default an only option in the current version being "Indicator", SVD performed on the weighted indicator matrix..}

\item{lvl.num}{logical value to specify the labels of the category levels, with default \code{TRUE} to indicate levels with numbers (V1, V2, etc.). Alternatively the level names are used.}
}
\value{
A list with the following components is available:
\item{Z}{Combined data frame of the sample and category level coordinates.}
\item{Zprin}{Sample coordinates (principal coordinates).}
\item{CLPstan}{Category level point coordinates (standard coordinates).}
\item{CLPnames}{Labels of the category level points as determined by \code{lvl.num} specification.}
\item{g}{the number of groups for MCA will always be 2.}
\item{g.names}{the group names refer to samples and variables in MCA}
\item{group.aes}{the vector of category levels for the grouping variable. This is to be used for \code{colour}, \code{pch} and \code{cex} specification.}
\item{dim.biplot}{The dimension of the biplot.}
\item{e.vects}{Depending on what was specified in \code{MCA} argument.}
}
\description{
This function produces a list of elements to be used for MCA biplot construction.
}
\examples{
biplot(as.data.frame(HairEyeColor)[,-4], center = FALSE) |> 
MCA(lvl.num = FALSE) |>  samples(col=c("purple3","forestgreen"), 
pch=c(15,17), label = TRUE, label.offset = 1, label.col="gray37") |> 
plot()
}
\seealso{
\code{\link[=biplot]{biplot()}}
}
