% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{means}
\alias{means}
\title{Format aesthetics for the class or group means}
\usage{
means (bp,  which = NULL, col = NULL, pch = 15, cex = 1, label = FALSE,
label.col = NULL,label.cex = 0.75, label.side = "bottom", label.offset = 0.5,
opacity = 1, shade.darker = TRUE)
}
\arguments{
\item{bp}{an object of class \code{biplot}.}

\item{which}{a vector containing the groups or classes for which the means should be displayed, with default \code{bp$g}.}

\item{col}{the colour(s) for the means, with default as the colour of the samples.}

\item{pch}{the plotting character(s) for the means, with default ■.}

\item{cex}{the character expansion(s) for the means, with default \code{1}.}

\item{label}{a logical value indicating whether the means should be labelled, with default \code{TRUE}.}

\item{label.col}{a vector of the same length as \code{which} with label colours for the means, with default as the colour of the means.}

\item{label.cex}{a vector of the same length as \code{which} with label text expansions for the means, with default \code{0.75}.}

\item{label.side}{the side at which the label of the plotted point appears, with default \code{bottom}. Note that unlike the argument \code{pos} in \code{text()}, options are "\code{bottom}", "\code{left}", "\code{top}", "\code{right}" and not \code{1}, \code{2}, \code{3}, \code{4}.}

\item{label.offset}{the offset of the label from the plotted point. See \code{?text} for a detailed explanation of the argument \code{offset}.}

\item{opacity}{transparency of means.}

\item{shade.darker}{automatically makes the colour of the means a darker shade than the default (or specified) colour}
}
\value{
A list with the following components is available:
\item{which}{which means to display.}
\item{col}{colour of the means.}
\item{pch}{plotting character of the means.}
\item{cex}{expansion of the plotting character of the means.}
\item{label}{logical, whether means should be labelled.}
\item{label.col}{colour of the label.}
\item{label.cex}{expansion of the label.}
\item{label.side}{side at which to plot the label of means.}
\item{label.offset}{offset of the label from the mean point.}
\item{opacity}{transparency of means}
}
\description{
This function allows the user to format the aesthetics for the class means or group means.
}
\details{
The number of classes or groups (defined by group.aes) is indicated as \code{g}. If an argument is not of length \code{g},
recycling is used.
}
\examples{
biplot(iris[,1:4]) |> PCA() |>
          means(col = "purple", pch = 15, cex = 2) |> plot()
}
