% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{PCA.biplot}
\alias{PCA.biplot}
\title{PCA biplot}
\usage{
\method{PCA}{biplot}(
  bp,
  dim.biplot = c(2, 1, 3),
  e.vects = 1:ncol(bp$X),
  group.aes = NULL,
  correlation.biplot = FALSE,
  ...
)
}
\arguments{
\item{bp}{An object of class \code{biplot} obtained from preceding function \code{biplot()}.}

\item{dim.biplot}{Dimension of the biplot. Only values 1, 2 and 3 are accepted, with default \code{2}.}

\item{e.vects}{Which eigenvectors (principal components) to extract, with default \code{1:dim.biplot}.}

\item{group.aes}{Optional argument. Vector of the same length as the number of rows in the data matrix
for differentiated aesthetics for samples.}

\item{correlation.biplot}{If \code{FALSE}, the distances between sample points are
optimally approximated in the biplot. If \code{TRUE}, the correlations between
variables are optimally approximated by the cosine of the angles between
axes. Default is \code{FALSE}.}
}
\value{
an object of class PCA
}
\description{
Performs calculations for a PCA biplot
}
\examples{
biplot(iris[,1:4]) |> PCA()

}
