% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{PCA}
\alias{PCA}
\title{PCA method}
\usage{
PCA(bp, e.vects = 1:2, ...)
}
\arguments{
\item{bp}{object of class biplot}

\item{e.vects}{which principal components to extract, defaults to first two}

\item{...}{more stuff}
}
\value{
object of class PCA
}
\description{
PCA method
}
