% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{plot.biplot}
\alias{plot.biplot}
\title{Plotting of the biplot}
\usage{
\method{plot}{biplot}(x, y = NULL, ...)
}
\arguments{
\item{x}{object of class biplot}

\item{y}{NULL}

\item{...}{more arguments}
}
\value{
biplot is created on graphics device
}
\description{
Plotting of the biplot
}
