% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.biplot}
\alias{plot.biplot}
\title{Generic Plotting function of objects of class biplot}
\usage{
\method{plot}{biplot}(
  x,
  exp.factor = 1.2,
  axis.predictivity = NULL,
  sample.predictivity = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{biplot}.}

\item{exp.factor}{numeric value with default axes of the biplot. Larger values for zooming out and
smaller values for zooming in with respect to sample points
in the biplot display.}

\item{axis.predictivity}{either logical or a numeric value between 0 and 1.
If it is a numeric value, this value is used as threshold so that
only axes with axis predictivity larger than the threshold is displayed.
If \code{axis.predictivity = TRUE}, the axis color is 'diluted' in
proportion with the axis predictivity.}

\item{sample.predictivity}{either a logical or a numeric value between 0 and 1.
If it is a numeric value, this value is used as threshold so that
only samples with sample predictivity larger than the threshold is displayed.
if \code{sample.predictivity = TRUE}, the sample size is shrinked in
proportion with the sample predictivity.}

\item{...}{additional arguments.}
}
\value{
An object of class \code{biplot}.
}
\description{
Generic Plotting function of objects of class biplot
}
\examples{
biplot (iris[,1:4]) |> PCA() |> plot()
}
