% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{plot.biplot}
\alias{plot.biplot}
\title{Generic Plotting function of objects of class biplot}
\usage{
\method{plot}{biplot}(x, exp.factor = 1.2, ...)
}
\arguments{
\item{x}{An object of class \code{biplot}.}

\item{exp.factor}{factor to expand plotting area beyond samples}

\item{...}{additional arguments}
}
\description{
Generic Plotting function of objects of class biplot
}
\examples{
biplot (iris[,1:4]) |> PCA() |> plot()
}
