% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{print.biplot}
\alias{print.biplot}
\title{Generic print function of objects of class biplot}
\usage{
\method{print}{biplot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{biplot}.}

\item{...}{additional arguments}
}
\description{
Generic print function of objects of class biplot
}
\examples{
out <- biplot (iris[,1:4]) |> PCA() |> plot()
out
}
