% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{reflect}
\alias{reflect}
\title{Reflect}
\usage{
reflect(bp, reflect.axis = c("FALSE", "x", "y", "xy"))
}
\arguments{
\item{bp}{object of class \code{biplot}}

\item{reflect.axis}{Defaults to FALSE. Possible values are "x" for reflection about x-axis or "y" for reflection about y-axis.}
}
\value{
An object of class \code{biplot}
}
\description{
Reflect
}
\examples{
biplot(iris[,1:4],group.aes = iris[,5]) |> PCA() |> reflect("x") |> plot()
biplot(iris[,1:4],group.aes = iris[,5]) |> PCA() |> reflect("y") |> plot()
biplot(iris[,1:4],group.aes = iris[,5]) |> PCA() |> reflect("xy") |> plot()
}
