% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{samples}
\alias{samples}
\title{Aesthetics for biplot samples}
\usage{
samples(
  bp,
  col = ez.col,
  pch = 3,
  cex = 1,
  label = F,
  label.cex = 0.75,
  label.side = "bottom",
  connected = F,
  alpha = 1
)
}
\arguments{
\item{bp}{biplot object}

\item{col}{sample colour}

\item{pch}{sample plotting character}

\item{cex}{sample character expansion}

\item{label}{logical, whether samples should be labelled or not}

\item{label.cex}{label text expansion}

\item{label.side}{side of the plotting character where label appears}

\item{connected}{logical, whether samples are connected in order of rows of data matrix}

\item{alpha}{opacity}
}
\value{
A list with the following components
{
\item{col}{Colour of the sample}
\item{pch}{Plotting character of the sample}
\item{cex}{Expansion of the plotting character}
\item{label}{Logical parameter whether labels appear}
\item{label.cex}{}
\item{label.side}{}
\item{connected}{}
\item{alpha}{}
\item{g}{number of groups}
}
}
\description{
Formatting of samples in biplot
}
\examples{
biplot(iris[,1:4]) |> PCA() |> samples(col="purple",pch=15) |> plot()
}
