% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{samples}
\alias{samples}
\title{Formatting of samples in biplot}
\usage{
samples(
  bp,
  col = c("blue", "green", "gold", "cyan", "magenta", "black", "red", "grey", "purple",
    "salmon"),
  pch = 3,
  cex = 1,
  label = F,
  label.cex = 0.75,
  label.side = "bottom",
  connected = F,
  alpha = 1
)
}
\arguments{
\item{bp}{biplot objects}

\item{col}{sample colour}

\item{pch}{sample plotting character}

\item{cex}{sample character expansion}

\item{label}{logical, whether samples should be labelled or not}

\item{label.cex}{label character expansion}

\item{label.side}{side of the plotting character where label appears}

\item{connected}{logical, whether samples are connected in order of rows of data matrix}

\item{alpha}{opacity}
}
\description{
Formatting of samples in biplot
}
