% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{samples}
\alias{samples}
\title{Aesthetics for biplot samples}
\usage{
samples (bp,  col = ez.col, pch = 3, cex = 1, label = FALSE,
label.cex = 0.75, label.side = "bottom", connected = FALSE, alpha = 1)
}
\arguments{
\item{bp}{Object of class biplot obtained from preceding function \code{biplot()}.}

\item{col}{Sample colour, with default \code{blue}.}

\item{pch}{Sample plotting character, with default \code{+}.}

\item{cex}{Sample character expansion, with default \code{1}.}

\item{label}{Logical argument, whether samples should be labelled or not, with default \code{FALSE}.}

\item{label.cex}{Label text expansion, with default \code{0.75}.}

\item{label.side}{Side of the plotting character where label appears, with default \code{bottom}.}

\item{connected}{Logical argument, whether samples are connected in order of rows of data matrix, with default \code{FALSE}.}

\item{alpha}{Opacity of sample plotting character, default is \code{1}.}
}
\value{
A list with the following components is available:
\item{col}{Colour of the samples.}
\item{pch}{Plotting character of the samples.}
\item{cex}{Expansion of the plotting character of the samples.}
\item{label}{TRUE or FALSE, as specified.}
\item{label.cex}{Expansion of the label.}
\item{label.side}{Side at which to plot the label of samples.}
\item{connected}{TRUE or FALSE, as specified.}
\item{alpha}{Opacity of the samples.}
\item{g}{Number of groups.}
}
\description{
This function allows formatting changes to samples.
}
\examples{
biplot(iris[,1:4]) |> PCA() |> samples(col="purple",pch=15) |> plot()
}
